Move IrisSessionPath and iris related NavigationController extension from Iris module to Navigation module
