--- ArtemisKit/Sources/Iris/Navigation/NavigationController+Iris.swift
@@ -1,12 +1,0 @@
-import Foundation
-import Navigation
-import SwiftUI
-
-public extension NavigationController {
-    func goToIrisSession(courseId: Int, sessionId: Int) {
-        goToCourse(id: courseId)
-        courseTab = .iris
-        selectedPath = IrisSessionPath(sessionId: sessionId, coursePath: CoursePath(id: courseId))
-        tabPath = NavigationPath()
-    }
-}
